unwind: support prelinked shared libraries
The code as previously written only works when the library is linked at an
offset of 0.  This is true for most shared objects, but not for prelinked
shared objects.  Need to set segbase to the difference between where the
library was actually loaded and where it expected to be loaded.