proc: make signal handler async-signal-safe
Calls to ptrace() and exit() are both not safe to be called from signal
handlers. The ptrace() calls are actually not needed either as simply
exiting will cause ptrace to detach. async-signal-safe calls are listed
in the "man 7 signal".

With the removal of the ptrace detach however we now need to send
SIGCONT even if we didn't send SIGSTOP because the automatic ptrace
detach that occurs will leave the process in the stopped state, even
though it is no longer being traced.

These issues were found when iterating tbstack over a large number of
processes with a timeout that would send SIGTERM if the whole process
took too long. Killing tbstack with SIGTERM would occasionally leave
processes in a stopped state. With these changes the SIGTERM
consistently leaves processes in the correct running state.