find_proc_info: subtract file offset from eh_frame map base
segbase and table_data are file offsets. The region is the virtual
memory region that contains the current ip. To convert the file offsets
to memory offsets, we need to add the load address of the start of the
text segment. (This assumes the eh_frame is part of that segment or
otherwise at the same displacement from the text in memory as in the
file.)

Adding region->start assumes that this region starts at the beginning of
the first segment and its file offset is 0. If the map is modified at
runtime by the program, as can happen when calling mlock() over a subset
of a region, the ip may be in a subdivided region and region->start
might not correspond to file offset 0. Luckily, we have the file offset
corresponding to each region, so we can simply subtract this from the
region start to get the file to memory offset. Usually region->offset
will be zero as previously assumed which is why this worked in most
cases.